+++ src/db/schema/index.ts
@@ -35,3 +35,4 @@ export * from './commerce';
 export * from './shipping-rates';
 export * from './book-reviews';
 export * from './book-author-bios';
+export * from './author-bios';